requirements.txt
django
djangorestframework
djangorestframework-simplejwt
gunicorn 
psycopg2-binary 
dj-database-url 
python-dotenv 
whitenoise
requests
django-cors-headers